chat function + assignment function

+++ services/course-service/src/main/java/com/example/courseservice/services/CourseSectionsService.java
@@ -5,12 +5,17 @@ import com.example.courseservice.dto.CourseSectionsDTO;
 import com.example.courseservice.dto.ResultDTO;
 import com.example.courseservice.entity.CourseEntity;
 import com.example.courseservice.entity.CourseSectionsEntity;
+import com.example.courseservice.entity.UserEnrolmentsEntity;
+import com.example.courseservice.enums.CourseRole;
 import com.example.courseservice.mapper.CourseMapper;
 import com.example.courseservice.mapper.CourseSectionsMapper;
 import com.example.courseservice.repository.CourseRepository;
 import com.example.courseservice.repository.CourseSectionsRepository;
+import com.example.courseservice.repository.EnrolRepository;
+import com.example.courseservice.repository.UserEnrolmentsRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -29,6 +34,12 @@ public class CourseSectionsService {
     CourseRepository courseRepository;
 
     @Autowired
+    private EnrolRepository enrolRepository;
+
+    @Autowired
+    private UserEnrolmentsRepository userEnrolmentsRepository;
+
+    @Autowired
     CourseSectionsMapper courseSectionsMapper;
 
     public ResponseEntity<ResultDTO> getAllCourseSectionsByCourseId(Long courseId){
@@ -40,7 +51,8 @@ public class CourseSectionsService {
     }
 
 
-    public ResponseEntity<ResultDTO> addCourseSection(CourseSectionsDTO courseSectionsDTO){
+    public ResponseEntity<ResultDTO> addCourseSection(CourseSectionsDTO courseSectionsDTO,
+                                                      String userId){
         ResultDTO resultDTO = new ResultDTO();
 
         if (courseSectionsDTO == null) {
@@ -61,6 +73,30 @@ public class CourseSectionsService {
             return ResponseEntity.badRequest().body(resultDTO);
         }
 
+        List<UserEnrolmentsEntity> userEnrolmentsEntityList = userEnrolmentsRepository.getAllUserEnrolmentsByCourseId(courseEntity.getId());
+        if (userEnrolmentsEntityList.isEmpty()) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("User not found");
+            return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+        }
+
+        if (!userEnrolmentsEntityList.stream().anyMatch(userEnrolmentsEntity -> userEnrolmentsEntity.getUserId().equals(userId))) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("User not enrolled in course");
+            return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+        }
+        // validate role teacher can create Assignmet
+
+        for (UserEnrolmentsEntity userEnrolments : userEnrolmentsEntityList) {
+            if (userEnrolments.getUserId().equals(userId)) {
+                if(userEnrolments.getEnrol().getCourseRole() == CourseRole.STUDENT){
+                    resultDTO.setStatus(0);
+                    resultDTO.setMessage("User dont have permission to create Course Section");
+                    return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+                }
+            }
+        }
+
         CourseSectionsEntity courseSectionsEntity = new CourseSectionsEntity();
         courseSectionsEntity.setName(courseSectionsDTO.getName());
         courseSectionsEntity.setSummary(courseSectionsDTO.getSummary());
@@ -75,13 +111,97 @@ public class CourseSectionsService {
         return ResponseEntity.ok(resultDTO);
     }
 
-    public ResponseEntity<ResultDTO> deleteCourseSectionById(Long id){
+    public ResponseEntity<ResultDTO> updateCourseSection(CourseSectionsDTO courseSectionsDTO, String userId){
+        ResultDTO resultDTO = new ResultDTO();
+        if (courseSectionsDTO == null) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("CourseSectionsDTO cannot be null");
+            return ResponseEntity.badRequest().body(resultDTO);
+        }
+
+        CourseSectionsEntity courseSectionsEntity = courseSectionsRepository.findById(courseSectionsDTO.getId()).orElse(null);
+        if (courseSectionsEntity == null) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("Course section not found");
+            return ResponseEntity.badRequest().body(resultDTO);
+        }
+
+        CourseEntity courseEntity = courseSectionsEntity.getCourse();
+
+        List<UserEnrolmentsEntity> userEnrolmentsEntityList = userEnrolmentsRepository.getAllUserEnrolmentsByCourseId(courseEntity.getId());
+        if (userEnrolmentsEntityList.isEmpty()) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("User not found");
+            return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+        }
+
+        if (!userEnrolmentsEntityList.stream().anyMatch(userEnrolmentsEntity -> userEnrolmentsEntity.getUserId().equals(userId))) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("User not enrolled in course");
+            return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+        }
+        // validate role teacher can create Assignmet
+
+        for (UserEnrolmentsEntity userEnrolments : userEnrolmentsEntityList) {
+            if (userEnrolments.getUserId().equals(userId)) {
+                if(userEnrolments.getEnrol().getCourseRole() == CourseRole.STUDENT){
+                    resultDTO.setStatus(0);
+                    resultDTO.setMessage("User dont have permission to create Course Section");
+                    return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+                }
+            }
+        }
+
+        courseSectionsEntity.setSummary(courseSectionsDTO.getSummary());
+        courseSectionsEntity = courseSectionsRepository.save(courseSectionsEntity);
+        resultDTO.setData(courseSectionsEntity);
+        resultDTO.setStatus(1);
+        return ResponseEntity.ok(resultDTO);
+    }
+
+
+    public ResponseEntity<ResultDTO> deleteCourseSectionById(Long id, String userId){
         ResultDTO resultDTO = new ResultDTO();
         Optional<CourseSectionsEntity> courseSectionsEntityOptional = courseSectionsRepository.findById(id);
         if(courseSectionsEntityOptional.isEmpty()){
             resultDTO.setStatus(2);
             resultDTO.setMessage("No data");
         }
+
+        CourseSectionsEntity courseSectionsEntity = courseSectionsEntityOptional.get();
+        CourseEntity courseEntity = courseSectionsEntity.getCourse();
+
+        if (courseEntity == null) {
+            resultDTO.setStatus(2);
+            resultDTO.setMessage("Course not found");
+            return ResponseEntity.badRequest().body(resultDTO);
+        }
+
+        List<UserEnrolmentsEntity> userEnrolmentsEntityList = userEnrolmentsRepository.getAllUserEnrolmentsByCourseId(courseEntity.getId());
+        if (userEnrolmentsEntityList.isEmpty()) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("User not found");
+            return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+        }
+
+        if (!userEnrolmentsEntityList.stream().anyMatch(userEnrolmentsEntity -> userEnrolmentsEntity.getUserId().equals(userId))) {
+            resultDTO.setStatus(0);
+            resultDTO.setMessage("User not enrolled in course");
+            return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+        }
+        // validate role teacher can create Assignmet
+
+        for (UserEnrolmentsEntity userEnrolments : userEnrolmentsEntityList) {
+            if (userEnrolments.getUserId().equals(userId)) {
+                if(userEnrolments.getEnrol().getCourseRole() == CourseRole.STUDENT){
+                    resultDTO.setStatus(0);
+                    resultDTO.setMessage("User dont have permission to create Course Section");
+                    return new ResponseEntity<>(resultDTO, HttpStatus.BAD_REQUEST);
+                }
+            }
+        }
+
+
         courseSectionsRepository.deleteById(id);
         resultDTO.setStatus(1);
         resultDTO.setMessage("Success");
